Refactor import paths for SettingsDialog to follow new directory structure
