+++ ui/lib/components/app_bar/settings_dialog.dart
