+++ ui/test/pages/home_page_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 import 'package:ui/pages/home_page.dart';
 import 'package:ui/app_state.dart';
-import 'package:ui/components/settings_dialog.dart';
+import 'package:ui/components/app_bar/settings_dialog.dart';
 import 'package:ui/helpers/http_helper.dart';
 import 'package:http/http.dart' as http;
 import 'package:mockito/annotations.dart';
